+++ app/models/season.rb
@@ -5,6 +5,10 @@ class Season < ApplicationRecord
   has_many :episodes, dependent: :destroy
   before_save -> { self.slug = name&.gsub(/[^a-z0-9\s]/i, "")&.parameterize }
 
+  scope :aired_in, lambda { |year|
+    where("EXTRACT(year FROM air_date) = ?", year)
+  }
+
   def poster
     Poster.new(tmdb_poster_path.presence || show.tmdb_poster_path)
   end